feat(billing): export slider props and components for use in getsentry
