+++ static/app/components/forms/controls/rangeSlider/sliderLabel.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const SliderLabel = styled('label')`
+  font-size: 14px;
+  margin-bottom: ${p => p.theme.grid}px;
+  color: ${p => p.theme.subText};
+`;
+
+export default SliderLabel;
